Add files via upload
This script would allow you to label and annotate the downloaded images.  The other objective of this project is also to label or create your own dataset for training. Any images dataset scraped from a website can used for labeling. When annotating our dataset, your screen will display a digit on your screen. You need to press the corresponding key on your keyboard to label the image and save the image to your disk.